Replace BufferedTexture with AsyncTextureReader
Replaces the BufferedTexture class and its usage with the new AsyncTextureReader class for asynchronous GPU texture reading. Updates all relevant includes, class members, and test cases to use AsyncTextureReader. Removes BufferedTexture implementation and its tests, adds new tests for AsyncTextureReader, and updates CMakeLists accordingly. Minor refactoring and cleanup in affected files.
