+++ src/DrawingEffect.hpp
@@ -116,7 +116,8 @@ public:
 				 gs_texture_t *targetIntermediateTexture, gs_texture_t *sourceTexture) noexcept;
 
 	void drawFinalImage(gs_texture_t *targetTexture, gs_texture_t *sourceTexture) noexcept;
-	void drawFinalImage(uint32_t width, uint32_t height, gs_texture_t *targetTexture, gs_texture_t *sourceTexture) noexcept;
+	void drawFinalImage(uint32_t width, uint32_t height, gs_texture_t *targetTexture,
+			    gs_texture_t *sourceTexture) noexcept;
 
 private:
 	void applyEffectPass(gs_technique_t *technique, gs_texture_t *sourceTexture) noexcept;
